Prevent rendering of HTML options
When there is no pre-existing selection, the previous
behaviour was to render all available selections to HTML.
This caused issues and could cause loading issues in the
future.

This commit serves to render an empty collection if there
is no currently-selected record and allow the selections to
be searched and inputted via ajax like it should.

Change-Id: Ie53674d23f6f2053f91fa278c38753c22b9011bc